+++ admin/include/view_post.php
@@ -31,7 +31,7 @@
                             echo"<td>{$post_content}</td>";
                             echo"<td>{$post_tags}</td>";
                             echo"<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                            echo"<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                            echo"<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this post?');\" href='posts.php?delete={$post_id}'>Delete</a></td>";
                              echo"</tr>";
                             
                         }
@@ -42,6 +42,10 @@
 <?php 
                      if(isset($_GET['delete']))
                         {
+                             if(isset($_SESSION['user_role']))
+                                {
+                                    if($_SESSION['user_role']==='admin')
+                                    {
                             $post_id=$_GET['delete'];
                                $query="DELETE FROM posts ";
                                 $query .="WHERE post_id=$post_id";   
@@ -51,6 +55,8 @@
                                 }
                             header("Location:posts.php");
                             
+                                  }
+                                }
                         }
-?>
+?>
