+++ admin/include/view_users.php
@@ -35,7 +35,7 @@
                             echo"<td><a href='users.php?change_role_to_admin={$user_id}'>Admin</a></td>";
                             echo"<td><a href='users.php?change_role_to_subscriber={$user_id}'>Subscriber</a></td>";
                             echo"<td><a href='users.php?source=edit_user&user_id={$user_id}'>Edit</a></td>";
-                            echo"<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
+                            echo"<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this user?');\" href='users.php?delete={$user_id}'>Delete</a></td>";
                              echo"</tr>";
                             
                         }
@@ -46,15 +46,20 @@
 <?php 
                      if(isset($_GET['delete']))
                         {
+                            if(isset($_SESSION['user_role']))
+                                {
+                                    if($_SESSION['user_role']==='admin')
+                                    {
                             $user_id=$_GET['delete'];
                                $query="DELETE FROM users ";
                                 $query .="WHERE user_id=$user_id";   
                                 $result=mysqli_query($connection,$query);
                                 if(!$result){
-                                    die("query failed" .mysqli_error($connection));
-                                }
+                                     die("query failed" .mysqli_error($connection));
+                                      }
                             header("Location:users.php");
-                            
+                                    }
+                                }
                         }
